pyproject.toml
[project]
name = "TimezoneBot"
version = "1.0.0"
description = "A Discord bot for keeping track of people's timezones"
authors = [{name="M3z0id"}]
requires-python = ">=3.13"

dependencies = [
    "py-cord==2.6.1",
    "mariadb==1.1.13",
    "datetime==5.5",
    "asyncio==4.0.0",
    "pycryptodome==3.23.0",
    "geoip2==5.1.0",
    "pytz==2025.2",
    "requests==2.32.5",
    "prompt_toolkit==3.0.51",
    "tzlocal==5.3.1",
    "sqlite3-to-mysql==2.4.2",
    "mysql-to-sqlite3==2.4.2",
    "dataclasses-json==0.6.7",
    "aiosqlite==0.21.0",
    "aiomysql==0.2.0",
    "aiofiles==24.1.0",
    "numpy==2.3.2",
    "matplotlib==3.10.5",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "packetsender.py"
]
target-version = "py313"
line-length = 150
indent-width = 4

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D",        # Docstring
    "ANN002",   # *args type annotation
    "ANN003",   # **kwargs type annotation
    "C901",     # function too complex
    "COM812",   # Trailing comma
    "DTZ005",   # datetime without timezone
    "ERA001",   # Commented out code
    "EM102",    # error message with no variables
    "FBT003",   # Boolean positional argument
    "INP001",   # __init__.py in packages
    "N802",     # lowercase function names
    "N803",     # lowercase arg names
    "N805",     # self in member functions
    "N806",     # lowercase var names
    "N815",     # mixCaseOmg
    "N816",     # camelCase
    "PLR0912",  # Too many branches
    "PGH004",   # Blanket noqa
    "PTH123",   # Path.open()
    "RUF006",   # Storing return value of a function
    "RUF012",   # Annotation with @typing.ClassVar
    "S104",     # Binding to all interfaces
    "S112",     # Logging exceptions
    "S606",     # Starting process without a shell
    "SLF001",   # Private member accessed
    "TRY003",   # Long exception message
    "TRY300",   # Use else block instead of catch
]
fixable = ["ALL"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
